Restructure sequence package.
Break piece sequences into a piece cache and a get-next-tetromino selector. Add seed values to the constructors of the randomized selectors.

There may be a need to possible re-seed the values (if we want consecutive games for the AI to be deterministic, for example) after the program starts. Also, we should be able to switch between sequences on-the-fly.